requirements.txt
asgiref==3.5.0
cycler==0.11.0
Django==4.0.3
fonttools==4.29.1
greenlet==1.1.2
kiwisolver==1.3.2
matplotlib==3.5.1
mysql-connector-python==8.0.28
mysqlclient==2.1.0
numpy==1.22.2
packaging==21.3
pandas==1.4.1
Pillow==9.0.1
protobuf==3.19.4
psycopg2==2.9.3
pyparsing==3.0.7
python-dateutil==2.8.2
pytz==2021.3
six==1.16.0
SQLAlchemy==1.4.31
sqlparse==0.4.2
tk==0.1.0
tzdata==2021.5
